feat(pySD): seperate reading of config file from cxx2py
