+++ docs/source/pySD/readconfigfile.rst
@@ -1,0 +1,5 @@
+READCONFIGFILE
+==============
+
+.. automodule:: pySD.readconfigfile
+   :members:
